requirements.txt
google-cloud-resource-manager
google-cloud-bigquery